+++ app/Livewire/Attendance/PollEdit.php
@@ -26,9 +26,9 @@ class PollEdit extends Component
     {
         $this->pollForm->update();
 
-        Log::info("Poll edited", [auth()->user(), $this->pollForm->poll]);
+        Log::info('Poll edited', [auth()->user(), $this->pollForm->poll]);
         NotificationMessage::addNotificationMessage(
-            new Item( __('The attendance poll has been successfully updated.'), ItemType::SUCCESS));
+            new Item(__('The attendance poll has been successfully updated.'), ItemType::SUCCESS));
 
         return redirect($this->previousUrl);
     }
@@ -37,9 +37,9 @@ class PollEdit extends Component
     {
         $this->pollForm->delete();
 
-        Log::info("Poll deleted", [auth()->user(), $this->pollForm->poll]);
+        Log::info('Poll deleted', [auth()->user(), $this->pollForm->poll]);
         NotificationMessage::addNotificationMessage(
-            new Item( __('The event has been successfully deleted.'), ItemType::WARNING));
+            new Item(__('The event has been successfully deleted.'), ItemType::WARNING));
 
         return redirect($this->previousUrl);
     }