+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## Unpublished
 ### Added
 - imprint and privacy policy page, editable via settings
+- flag for public poll statistic, if checked poll stats can be shown to everyone with link
 
 ## v0.7.10 - 2024-05-05
 ### Added